refactor(pricing): give the usable-rate question a module of its own
`is_usable_rate` lived in `payment/models.py`, which pulls in httpx, fastapi,
the DB session and settings. Every consumer therefore reached it through a
deferred import to avoid the cycle: five of them, in four files.

Move the predicate and the billable-rate tuple to `payment/rates.py`, which
imports only `math`, and let the call sites import it normally. The rationale
now lives in that one docstring, so each guard says what it does with the
answer instead of repeating why the answer matters.

No behaviour change.
